Refactor t2
