+++ run_t2.py
@@ -1,6 +1,5 @@
 # change the import to the code you want to run
-from rpi.task2_indoor import TaskTwo
+from rpi.task2_indoor import TaskTwoIndoor
 
 if __name__ == "__main__":
-    bot = TaskTwo()
-    bot.start()
+    TaskTwoIndoor.start()
